+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- **PODAAC-2541**
+  - a build script under /builder directory, and a jenkins job to build and push release to public github.
 ## [1.4.1] - 2020-08-17
 ### Added
 ### changed